[JENKINS-52964] Add SCMFileSystem.supports(SCMSourceDescriptor)
It's a pain that we can't tell whether a given `SCMSource`
implementation can support an `SCMFileSystem` without having an actual
instance of the `SCMSource` - that makes it impossible, for example,
to filter a list of possible `SCMSource`s in a configuration UI to
just those which support `SCMFileSystem`.

This is a proposed way to address that - add
`SCMFileSystem.Builder#supports(SCMSourceDescriptor)`, and
`SCMFileSystem#supports(SCMSourceDescriptor)`. The `Builder` method
will return false unless overridden, to preserve compatibility, and
it's up to individual `Builder` implementations to decide whether they
want to say "yes, the corresponding `SCMSource` for this
`SCMSourceDescriptor` is supported" without an actual instance to
check against for more details.